refactor: move nvim stuff into packages
