+++ nvim/lua/treesitter.lua
@@ -1,0 +1,9 @@
+require('nvim-treesitter.configs').setup({
+  highlight = {
+    enable = true,
+  },
+  ensure_installed = 'maintained',
+  indent = {
+    enable = true,
+  },
+})
